Add files via upload
Create an associative array where the key of each element is the name of a Simpson family member (Homer, Marge, Bart, Lisa, Maggie) and the value of each element is the person's age (Homer is 38, Marge is 36, Bart is 10, Lisa is 8, and Maggie is 1). For each element in the array, generate this line of output followed by a line break: "<person's name> is <person's age> years old".