Add files via upload
In JavaScript, associative arrays and objects are actually the same thing. You can use either "associative array" syntax or object syntax. The keys of an associative array are the properties of an object. Use object syntax to create a few 2-letter state abbreviations and their state names. Loop through the properties of the object to generate a checkbox for each state where the value of the checkbox is the state abbreviation and the label for the checkbox is the state name.