Add files via upload
Create an array that contains a few 2-letter state abbreviations and a second parallel array that contains the corresponding state names. For each element in the arrays, generate a checkbox where the value of the checkbox is the state abbreviation and the label for the checkbox is the state name.