Add files via upload
Write a function that takes 2 args. The 1st arg is an array and the 2nd arg is any value that could be a string, a number, or a boolean. Return the index of the value in the array. If the value occurs more than once in the array, return the index of the first occurrence. If the value is not in the array, return -1. Don't use array.indexOf(). Confirm that your function and array.indexOf() return the same results.