Add files via upload
Write a function that takes 2 args. The 1st arg is an array and the 2nd arg is any value that could be a string or a number. In the function, determine if the value (2nd arg) is equal to one of the elements in the array. If it is, return true. If it's not, return false.