Add files via upload
Write a function that takes 2 args, an array and a number. The 2nd arg represents a valid index of the array. Return a new array that contains the elements of the array passed in except the element at the index passed in. For example, if the array passed in has the elements 5,10,15,20,25 and the index passed in is 2, the returned array will have the elements 5,10,20,25. Do not use any array method.