Add files via upload
Write a function that takes 2 args. The 1st arg is a string with substrings separated by a delimiter character. For example, "buy groceries|do homework|pay bills". The 2nd arg is the delimiter character. Return an array that contains an element for each substring in the 1st arg. Don't use str.split(). Confirm that your function and str.split() return the same results.